fix all errors related to compiling with flags  Changes to be committed: 	modified:   Makefile 	modified:   WSNetworking/Headers/ConfigFileParser.hpp 	modified:   WSNetworking/Headers/RequestParser.hpp 	modified:   WSNetworking/Includes/Include.hpp 	modified:   WSNetworking/Sources/ConfigFileParser.cpp 	modified:   WSNetworking/Sources/ConfigLocationParser.cpp 	modified:   WSNetworking/Sources/ConfigServerParser.cpp 	modified:   WSNetworking/Sources/MainClient.cpp 	modified:   WSNetworking/Sources/MainServer.cpp 	modified:   WSNetworking/Sources/RequestParser.cpp

+++ WSNetworking/Headers/ConfigFileParser.hpp
@@ -30,11 +30,11 @@ class ConfigFileParser {
 
   private:
 	// Methods
-	void open_config_file();
-	void read_config_file();
-	int	 get_start_end_server(size_t pos, string delimiter);
-	void split_config_file();
-	void parse_config_file();
+	void   open_config_file();
+	void   read_config_file();
+	size_t get_start_end_server(size_t pos, string delimiter);
+	void   split_config_file();
+	void   parse_config_file();
 };
 
 #endif // CONFIGFILEPARSER_HPP
